+++ core/source/pkx/PK7.cpp
@@ -167,11 +167,11 @@ void PK7::experience(u32 v)
     *(u32*)(data + 0x10) = v;
 }
 
-u8 PK7::ability(void) const
+u16 PK7::ability(void) const
 {
     return data[0x14];
 }
-void PK7::ability(u8 v)
+void PK7::ability(u16 v)
 {
     data[0x14] = v;
 }
@@ -245,11 +245,11 @@ void PK7::gender(u8 v)
     data[0x1D] = u8((data[0x1D] & ~0x06) | (v << 1));
 }
 
-u8 PK7::alternativeForm(void) const
+u16 PK7::alternativeForm(void) const
 {
     return data[0x1D] >> 3;
 }
-void PK7::alternativeForm(u8 v)
+void PK7::alternativeForm(u16 v)
 {
     data[0x1D] = u8((data[0x1D] & 0x07) | (v << 3));
 }
